src/app/product/product.component.spec.ts
/* tslint:disable:no-unused-variable */
import { async, ComponentFixture, TestBed } from '@angular/core/testing';
import { By } from '@angular/platform-browser';
import { DebugElement } from '@angular/core';
import { FormsModule } from '@angular/forms';

import { ProductComponent } from './product.component';

describe('ProductComponent', () => {
  let component: ProductComponent;
  let fixture: ComponentFixture<ProductComponent>;

  beforeEach(async(() => {
    TestBed.configureTestingModule({
      declarations: [ ProductComponent ],
      imports:[FormsModule]
    })
    .compileComponents();
  }));

  beforeEach(() => {
    fixture = TestBed.createComponent(ProductComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('should correctly calculate the bmi', () => {
    component.height = 165;
    component.weight = 65;

    component.bmi();
   fixture.detectChanges();

    expect(component.result).toBe(23);

  });

  it('should correctly calculate the bmi', () => {
    component.height = 0;
    component.weight = 0;

    component.bmi();
   fixture.detectChanges();

    expect(component.result).toBe(0);

  });

});




